pyproject.toml
[project]
name = "smart-qna"
version = "0.1.0"
description = "Smart question answering system using RAG"
authors = [
    {name = "Arnab Halder"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "langchain (>=0.3.22,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
    "google-cloud-storage (>=2.18,<3.0)",
    "langchain-google-vertexai (>=2.0.18,<3.0.0)",
    "pinecone (>=6.0.2,<7.0.0)",
    "langchain-pinecone (>=0.2.4,<0.3.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
jupyter = "^1.1.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
